split up auto_materialize_rule.py
