+++ python_modules/dagster/dagster/_core/scheduler/instigation.py
@@ -6,7 +6,9 @@ from typing_extensions import TypeAlias
 
 import dagster._check as check
 from dagster._core.definitions import RunRequest
-from dagster._core.definitions.auto_materialize_rule import AutoMaterializeAssetEvaluation
+from dagster._core.definitions.auto_materialize_rule_evaluation import (
+    AutoMaterializeAssetEvaluation,
+)
 from dagster._core.definitions.events import AssetKey, AssetKeyPartitionKey
 
 # re-export
