+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/scenarios/partition_scenarios.py
@@ -6,10 +6,8 @@ from dagster import (
     StaticPartitionsDefinition,
 )
 from dagster._core.definitions.auto_materialize_policy import AutoMaterializePolicy
-from dagster._core.definitions.auto_materialize_rule import (
-    AutoMaterializeRule,
-    AutoMaterializeRuleEvaluation,
-)
+from dagster._core.definitions.auto_materialize_rule import AutoMaterializeRule
+from dagster._core.definitions.auto_materialize_rule_evaluation import AutoMaterializeRuleEvaluation
 from dagster._core.definitions.partition import (
     DynamicPartitionsDefinition,
 )
